:recycle: move file utils/api/otlp to convert.go and unmarshal.go
